refactor: move the world coordinate helper type next to the projection that supplies it
