+++ src/geo/mercator_coordinate.test.ts
@@ -1,6 +1,6 @@
 import {describe, test, expect} from 'vitest';
 import {LngLat} from './lng_lat.ts';
-import {MercatorCoordinate, mercatorScale} from './mercator_coordinate.ts';
+import {MercatorCoordinate, mercatorScale, mercatorWorldCoordinates} from './mercator_coordinate.ts';
 
 describe('LngLat', () => {
     test('constructor', () => {
@@ -31,5 +31,28 @@ describe('LngLat', () => {
     test('meterInMercatorCoordinateUnits', () => {
         const nullIsland = new LngLat(0, 0);
         expect(MercatorCoordinate.fromLngLat(nullIsland).meterInMercatorCoordinateUnits()).toBe(2.4981121214570498e-8);
+    });
+});
+
+function createSamplePoints(): Array<[number, number]> {
+    return [
+        [0, 0],
+        [-180, -85],
+        [179.999, 85],
+        [12.5, 41.9],
+        [-122.4, 37.8],
+        [151.2, -33.9],
+        [37.6, 55.8],
+    ];
+}
+
+describe('mercatorWorldCoordinates', () => {
+    test('round-trips lng/lat through world coordinates', () => {
+        for (const [lng, lat] of createSamplePoints()) {
+            const {x, y} = mercatorWorldCoordinates.worldFromLngLat(lng, lat);
+            const back = mercatorWorldCoordinates.lngLatFromWorld(x, y);
+            expect(back.lng).toBeCloseTo(lng, 10);
+            expect(back.lat).toBeCloseTo(lat, 10);
+        }
     });
 });
