+++ src/geo/projection/globe_projection.ts
@@ -11,7 +11,8 @@ import {type SubdivisionGranularitySetting} from '../../render/subdivision_granu
 import {type Context} from '../../webgl/context.ts';
 import {type CanonicalTileID} from '../../tile/tile_id.ts';
 import {type Mesh} from '../../render/mesh.ts';
-import {mercatorWorldCoordinates, type WorldCoordinateHelper} from './world_coordinate_helper.ts';
+import type {WorldCoordinateHelper} from './projection.ts';
+import {mercatorWorldCoordinates} from '../mercator_coordinate.ts';
 
 export class GlobeProjection extends Evented implements Projection {
     properties: PossiblyEvaluated<ProjectionProps, ProjectionPropsPossiblyEvaluated>;
